♻️ (aa-auth): Move AuthError, Tenant, AuthenticatedCaller, scope, gate
Complete the extraction by moving the remaining auth framework into the
`aa-auth` leaf crate: the `scope`/`api_key`/`jwt`/`gate` modules (via `git mv`,
preserving history) and the core `AuthError`, `Tenant`, and
`AuthenticatedCaller` extractor (with its `FromRequestParts` impl and the
tenant-guard tests) into `aa-auth/src/lib.rs`. Internal `crate::auth::…`
references in the moved files are rewritten to the leaf's own crate root; the
deny-by-default gate keeps identical behavior.

`aa-api/src/auth/mod.rs` becomes a thin re-export facade
(`pub use aa_auth::{api_key, config, gate, jwt, rate_limit, scope, AuthError,
AuthenticatedCaller, Tenant}`), so all ~27 `crate::auth::…` call sites and the
`aa_api::auth::…` integration-test paths resolve unchanged. `policy_auth` stays
in `aa-api` because it couples to `aa_gateway::policy` and would re-introduce the
gateway dependency into the leaf. OpenAPI spec is byte-identical; behavior is
unchanged.

Refs AAASM-3904
Refs AAASM-3905
Refs AAASM-3906
