+++ aa-auth/src/gate.rs
@@ -5,9 +5,9 @@
 //! that forgot to do so (e.g. the alert-rule CRUD endpoints, AAASM-3129)
 //! silently shipped unauthenticated. This module provides a single
 //! deny-by-default [`require_authentication`] middleware that is applied as a
-//! `route_layer` over the protected sub-router in [`crate::routes`], so a new
-//! route is authenticated unless it is explicitly mounted on the public
-//! router.
+//! `route_layer` over the protected sub-router in the consuming service's
+//! router, so a new route is authenticated unless it is explicitly mounted on
+//! the public router.
 //!
 //! The gate reuses the existing [`AuthenticatedCaller`] extractor, so it
 //! honours `AuthMode::Off` (bypass) and the API-key / JWT validation and
@@ -18,7 +18,7 @@ use axum::middleware::Next;
 use axum::response::{IntoResponse, Response};
 use http::request::Parts;
 
-use crate::auth::AuthenticatedCaller;
+use crate::AuthenticatedCaller;
 
 /// Deny-by-default authentication middleware.
 ///
@@ -26,7 +26,7 @@ use crate::auth::AuthenticatedCaller;
 /// request proceeds to the inner service; on failure the [`AuthError`] is
 /// rendered (401 / 403 / 429) and the inner handler is never reached.
 ///
-/// [`AuthError`]: crate::auth::AuthError
+/// [`AuthError`]: crate::AuthError
 pub async fn require_authentication(request: Request, next: Next) -> Response {
     let (mut parts, body): (Parts, _) = request.into_parts();
